Rename error view, router mode passed in "History"
